icode

+++ leetcode/46_Permutations.h
@@ -18,7 +18,7 @@ public:
         }
         else
         {
-            for (int i = 0; i < nums.size(); i++)
+            for (int i = 0; i < (int)nums.size(); i++)
             {
                 if (find(item.begin(), item.end(), nums[i]) == item.end())
                 {
